[5116] Prevent custom appearances from being lost after exporting/importing a project
Bug: https://github.com/eclipse-sirius/sirius-web/issues/5116
