+++ CHANGELOG.adoc
@@ -105,6 +105,7 @@ Utility classes to help create assertions on those flux have been introduced to 
 - https://github.com/eclipse-sirius/sirius-web/issues/5107[#5107] [explorer] Provide the new `existingRepresentations` variable to the _Domain explorer by DSL_ explorer description.
 - https://github.com/eclipse-sirius/sirius-web/issues/5080[#5080] [diagram] Prevent border nodes from being resized during their parent resizing
 - https://github.com/eclipse-sirius/sirius-web/issues/5065[#5065] [diagram] Prevent adding bendpoints on a simple click
+- https://github.com/eclipse-sirius/sirius-web/issues/5116[#5116] [diagram] Prevent custom appearances from being lost after exporting/importing a project
 
 
 === New Features
